refactor: server renderer factory
